+++ src/CfdiUtils/Elements/Retenciones10/Emisor.php
@@ -1,0 +1,12 @@
+<?php
+namespace CfdiUtils\Elements\Retenciones10;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class Emisor extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'retenciones:Emisor';
+    }
+}