+++ src/CfdiUtils/Elements/Retenciones10/ImpRetenidos.php
@@ -1,0 +1,12 @@
+<?php
+namespace CfdiUtils\Elements\Retenciones10;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class ImpRetenidos extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'retenciones:ImpRetenidos';
+    }
+}