+++ src/CfdiUtils/Elements/Retenciones10/Nacional.php
@@ -1,0 +1,12 @@
+<?php
+namespace CfdiUtils\Elements\Retenciones10;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class Nacional extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'retenciones:Nacional';
+    }
+}