+++ tests/CfdiUtilsTests/Elements/Retenciones10/AddendaTest.php
@@ -1,0 +1,39 @@
+<?php
+namespace CfdiUtilsTests\Elements\Retenciones10;
+
+use CfdiUtils\Elements\Retenciones10\Addenda;
+use CfdiUtils\Nodes\Node;
+use PHPUnit\Framework\TestCase;
+
+class AddendaTest extends TestCase
+{
+    /** @var Addenda */
+    public $element;
+
+    public function setUp()
+    {
+        parent::setUp();
+        $this->element = new Addenda();
+    }
+
+    public function testGetElementName()
+    {
+        $this->assertSame('retenciones:Addenda', $this->element->getElementName());
+    }
+
+    public function testAdd()
+    {
+        $this->assertCount(0, $this->element);
+
+        $firstChild = new Node('first');
+        $addReturn = $this->element->add($firstChild);
+        $this->assertSame($this->element, $addReturn);
+        $this->assertCount(1, $this->element);
+        $this->assertSame($firstChild, $this->element->searchNode('first'));
+
+        $secondChild = new Node('second');
+        $this->element->add($secondChild);
+        $this->assertCount(2, $this->element);
+        $this->assertSame($secondChild, $this->element->searchNode('second'));
+    }
+}